fix: change int to long

+++ src/titular.h
@@ -4,7 +4,7 @@
 typedef struct 
 {
     char nombre [20];
-    int cuit;
+    long int cuit;
     char tipoDocumento [10];
     int nroDocumento; //[PK]
     char fechaNacimiento [20]; //Se puede usar DATE
